+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 ## Unreleased
 - Fix lazyness bug in `discover-network`: was sending multiple `Who-Is` simultaneously.
+- `boot-up!` returns the device ID.
 
 ## [1.2.0] - 2024-01-21
 - `with-temp-devices` no longer terminates and restarts the local device.